backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
langchain>=0.3.13
langchain-openai>=0.2.14
langgraph>=0.2.62
pydantic>=2.10.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
scikit-learn>=1.6.1
xgboost>=2.1.3
pandas>=2.2.3
numpy>=1.26.4,<2.0.0
simpy>=4.1.1
pulp>=2.9.0
joblib>=1.4.2
httpx>=0.28.1
sse-starlette>=2.2.1
python-multipart>=0.0.20
